+++ _P1/Logic/src/main/java/vdm/p1/logic/objects/ReadingTable.java
@@ -1,0 +1,20 @@
+package vdm.p1.logic.objects;
+
+import vdm.p1.engine.IFont;
+import vdm.p1.logic.objects.Cell;
+import vdm.p1.logic.objects.Table;
+import vdm.p1.logic.objects.Text;
+
+public class ReadingTable extends Table {
+	private Cell[][] cells;
+	private boolean[][] solutions;
+	private IFont font;
+	private int rows;
+	private int columns;
+
+	public ReadingTable(String filename, IFont font) {
+
+	}
+
+
+}
